% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_batch.R
\name{AUCell_batch}
\alias{AUCell_batch}
\title{Run AUCell Scoring in Batches:}
\usage{
AUCell_batch(inp_data, genesets, num_batches = 10, num_cores = 1)
}
\arguments{
\item{inp_data}{Cell x Gene expression matrix with cells as columns and genes as rows. Raw counts are acceptable.}

\item{genesets}{List of lists, which each component list being the genes that belong to an individual geneset to score by AUCell.}

\item{num_batches}{Number of batches to split the Cell x Gene expression matrix into. These will be processed sequentially to save memory.}

\item{num_cores}{Number of cores to use for parallel computation of AUCell scores.}
}
\value{
A matrix of AUCell scores for each geneset for each cell in the input dataset
}
\description{
Split Cell x Gene expression matrix into batches and run AUCell to reduce working memory load.
}
