% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ReferenceObject.R
\name{create_ReferenceObject}
\alias{create_ReferenceObject}
\title{Create Seurat Object from Symphony Reference. No Gene Expression data but can visualize annotations}
\usage{
create_ReferenceObject(ref_obj)
}
\arguments{
\item{ref_obj}{Custom Symphony Reference object}
}
\value{
A Seurat object with embeddings and metadata from the reference object. To save space, no raw count data is present in this object.
}
\description{
Create Seurat Object from Symphony Reference. No Gene Expression data but can visualize annotations
}
