% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Composition.R
\name{get_Composition}
\alias{get_Composition}
\title{Get Composition for projected CellTypes}
\usage{
get_Composition(
  query_obj,
  donor_key,
  celltype_label = "predicted_CellType",
  mapQC_col = "mapping_error_QC",
  knn_prob_cutoff = NULL,
  return_type = c("long", "count", "proportion")
)
}
\arguments{
\item{query_obj}{Query Seurat object as returned by Symphony map_Query}

\item{donor_key}{Metadata column specifying donor to tally celltype abundance by}

\item{celltype_label}{Column in query metadata with projected CellType classification.}

\item{mapQC_col}{Column in query object with mapping error QC status for each cell (either "Pass" or "Fail"). Can set as NULL if no filtering desired.}

\item{knn_prob_cutoff}{CellType classification KNN probability threshold for composition analysis. Default is NULL.}

\item{return_type}{Format to return composition data, either matrix with counts, matrix with proportions, or in a long format.}
}
\value{
A data frame with celltype abundance data for each donor.
}
\description{
Get Composition for projected CellTypes
}
