% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnPredict_Seurat.R
\name{knnPredict_Seurat}
\alias{knnPredict_Seurat}
\title{KNN Prediction of annotation label from reference to query
This is modified from the original Seurat_Utils.R script within the Symphony github page}
\usage{
knnPredict_Seurat(
  query_obj,
  ref_obj,
  label_transfer,
  col_name,
  k = 5,
  confidence = TRUE,
  seed = 0
)
}
\arguments{
\item{query_obj}{Query Seurat object as returned by map_Query}

\item{ref_obj}{Custom Symphony Reference object}

\item{label_transfer}{Annotation column from Symphony Reference object to use for classification}

\item{col_name}{Name of column in query object to store annotations}

\item{k}{Number of nearest neighbours from Symphony Reference to use for classification}

\item{confidence}{Whether to return KNN confidence scores (proportion of neighbours voting for the predicted annotation)}

\item{seed}{random seed (k-NN has some stochasticity in the case of ties)}
}
\value{
A Seurat object, with predicted reference labels stored in the 'col_name' column of the meta.data
}
\description{
KNN Prediction of annotation label from reference to query
This is modified from the original Seurat_Utils.R script within the Symphony github page
}
