% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Genesets_gmt.R
\name{load_Genesets_gmt}
\alias{load_Genesets_gmt}
\title{Load Geneset:}
\usage{
load_Genesets_gmt(path, ignore_cols = 2)
}
\arguments{
\item{path}{path to geneset file, should end with .gmt}

\item{ignore_cols}{Number of columns to skip from gmt file for extracting genes. First two columns are typically geneset name and description, so these will be skipped when extracting a list of genes for each geneset.}
}
\value{
A list of lists, with each component list being the genes belonging to an individual geneset.
}
\description{
Loads genesets stored in a gmt file into R as a list of lists.
}
