% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_Query.R
\name{map_Query}
\alias{map_Query}
\title{Function for mapping query data to Bone Marrow Reference
This is modified from the original Seurat_Utils.R function script within the Symphony github page}
\usage{
map_Query(
  exp_query,
  metadata_query,
  ref_obj,
  vars = NULL,
  verbose = TRUE,
  do_normalize = TRUE,
  do_umap = TRUE,
  sigma = 0.1
)
}
\arguments{
\item{exp_query}{Query gene expression (genes by cells). Raw counts are preferred.}

\item{metadata_query}{Query metadata (cells by attributes)}

\item{ref_obj}{Single cell hematopoiesis reference}

\item{vars}{Query batch variable(s) to integrate over (column names in metadata)}

\item{verbose}{Verbose output}

\item{do_normalize}{Perform log(CP10K+1) normalization on query expression. Default is TRUE}

\item{do_umap}{Perform umap projection into reference UMAP (if reference includes a uwot model). Default is TRUE}

\item{sigma}{Fuzziness parameter for soft clustering (sigma = 1 is hard clustering). Default is 0.1}
}
\value{
A Seurat object with mapped harmony and umap embeddings. Expression data is in the SymphonyQuery
}
\description{
Function for mapping query data to Bone Marrow Reference
This is modified from the original Seurat_Utils.R function script within the Symphony github page
}
