% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MappingErrorQC.R
\name{plot_MappingErrorQC}
\alias{plot_MappingErrorQC}
\title{Visualize Mapping error cutoff for QC purposes.
Low quality cells with low RNA counts and low transcriptional diversity tend to erroneously map to orthochromatic erythroblasts.
This will help the user to set a mapping error QC threshold to remove these low quality cells}
\usage{
plot_MappingErrorQC(
  query,
  mapQC_values = "mapping_error_score",
  mapQC_class = "mapping_error_QC"
)
}
\arguments{
\item{query}{Query Seurat object after symphony mapping and mapping error QC calculation}

\item{mapQC_values}{Column in query object with mapping error QC status for each cell (either "Pass" or "Fail"). Can set as NULL if no filtering desired.}

\item{mapQC_class}{Column in query object with mapping error QC status for each cell (either "Pass" or "Fail"). Can set as NULL if no filtering desired.}
}
\value{
Diagnostic QC plots to evaluate the current mapping error QC cutoff.
}
\description{
Visualize Mapping error cutoff for QC purposes.
Low quality cells with low RNA counts and low transcriptional diversity tend to erroneously map to orthochromatic erythroblasts.
This will help the user to set a mapping error QC threshold to remove these low quality cells
}
