% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Projection_BM.R
\name{plot_Projection_BM}
\alias{plot_Projection_BM}
\title{Density plots for bone marrow projection results}
\usage{
plot_Projection_BM(
  query_obj,
  batch_key,
  sample_name,
  ref_obj,
  Hierarchy_only = FALSE,
  downsample_reference = TRUE,
  downsample_frac = 0.25,
  query_point_size = 0.2,
  query_contour_size = 0.3,
  saveplot = TRUE,
  device = "pdf",
  save_folder = "projectionFigures/"
)
}
\arguments{
\item{query_obj}{Query Seurat object as returned by Symphony map_Query}

\item{batch_key}{Batch key specifying condition to subset and display samples by}

\item{sample_name}{Specific sample to display density plot for}

\item{ref_obj}{Bone Marrow Reference object to pull background UMAP plot from}

\item{Hierarchy_only}{Whether to only show the Hematopoietic hierarchy and exclude T/NK/Plasma/Stromal cells from the plot. Default is FALSE}

\item{downsample_reference}{Whether to downsample the BM reference to reduce file size. Default is TRUE}

\item{downsample_frac}{If downsampling reference, what proportion of cells to sample from the BM reference. Default is 0.25}

\item{query_point_size}{Point size of individual query cells on plot. Default is 0.2}

\item{query_contour_size}{Contour size depicting query cell density on plot. Default is 0.3}

\item{saveplot}{Whether to save the plot. Default is TRUE}

\item{device}{Whether to save the plot as a pdf or png. Default is pdf}

\item{save_folder}{Folder to save plots in. If this folder does not exist, it will be created}
}
\value{
A ggplot object with projected coordinates of the query cells overlaid over a background of the reference cells.
}
\description{
Density plots for bone marrow projection results
}
