% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Pseudotime.R
\name{predict_Pseudotime}
\alias{predict_Pseudotime}
\title{Annotate Hematopoietic Hierarchy Pseudotime metric}
\usage{
predict_Pseudotime(
  query_obj,
  ref_obj,
  pseudotime_label = "Pseudotime",
  k = 30,
  mapQC_class = "mapping_error_QC",
  initial_label = "initial_Pseudotime",
  final_label = "predicted_Pseudotime"
)
}
\arguments{
\item{query_obj}{Query Seurat object as returned by Symphony mapQuery}

\item{ref_obj}{Custom Symphony Reference object}

\item{pseudotime_label}{Pseudotime annotation column from Symphony Reference object}

\item{k}{Number of nearest neighbours from Symphony Reference to use for pseudotime annotation}

\item{mapQC_class}{Column in query object with mapping error QC status for each cell (either "Pass" or "Fail"). Can set as NULL if no filtering desired.}

\item{initial_label}{Column name to assign initial pseudotime annotations to, prior to mapQC filtering}

\item{final_label}{Column name to assign final pseudotime annotations to, following mapQC filtering}
}
\value{
A Seurat object, with annotated pseudotime labels stored in the 'final_label' column of the meta.data.
}
\description{
Assigns each query cell the median pseudotime of k-nearest neighbours by UMAP from the reference map.
Provides initial estimates for all cells regardless of mapping QC filter
Provides final estimates for mapQC passing cells and assigns mapQC failing cells as NA.
}
