% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ProjectionResults.R
\name{save_ProjectionResults}
\alias{save_ProjectionResults}
\title{Function to save CellType and Pseudotime predictions and UMAP coordinates}
\usage{
save_ProjectionResults(
  query_obj,
  file_name,
  celltype_label = "predicted_CellType",
  celltype_KNNprob_label = "predicted_CellType_prob",
  pseudotime_label = "predicted_Pseudotime",
  save_AUCell_scores = FALSE
)
}
\arguments{
\item{query_obj}{Query Seurat object after symphony mapping and celltype classification}

\item{file_name}{File name of csv to store projection annotations and UMAP coordinates}

\item{celltype_label}{Column in query metadata with projected CellType classification.}

\item{celltype_KNNprob_label}{Column in query metadata with KNN probabilities of the projected CellType classification.}

\item{pseudotime_label}{Column in query metadata with predicted Pseudotime values}

\item{save_AUCell_scores}{Boolean. Whether to save AUCell scores along with projection results. If TRUE, looks for metadata columns with suffix '_AUC' and saves those columns. This only works if AUCell scores are saved to the metadata. If you have them saved as an assay within the seurat object, please just save the seurat object with saveRDS.}
}
\value{
Saves a csv file with projected annotations to file_name
}
\description{
Function to save CellType and Pseudotime predictions and UMAP coordinates
}
