% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_Genesets_AUCell.R
\name{score_Genesets_AUCell}
\alias{score_Genesets_AUCell}
\title{Scoring Genesets on Seurat object by AUCell:}
\usage{
score_Genesets_AUCell(
  scrna,
  genesets,
  nbatches = 10,
  ncores = 1,
  output = c("metadata", "assay", "dataframe"),
  assay_name = "AUCell"
)
}
\arguments{
\item{scrna}{Seurat object to be scored by AUCell. "counts" slot within RNA assay will be used so this must not be empty.}

\item{genesets}{List of lists, which each component list being the genes that belong to an individual geneset to score by AUCell.}

\item{nbatches}{Number of batches to split the Cell x Gene expression matrix into. These will be processed sequentially to save memory.}

\item{ncores}{Number of cores to use for parallel computation of AUCell scores.}

\item{output}{Preferred output format of AUCell scores. This can either be "metadata", "assay", or "dataframe". "metadata" returns a seurat object with AUCell scores in the metadata. "assay" returns a seurat object with AUCell scores as a separate assay. "dataframe" returns a dataframe with AUCell scores for each cell.}

\item{assay_name}{If preferred output is "assay", specify the name of the assay to store the results in. Default is "AUCell".}
}
\value{
Depending on output parameter, either a seurat object with AUCell scores within metadata or a separate assay, or AUCell scores as a separate dataframe
}
\description{
Split Cell x Gene expression matrix into batches and run AUCell to reduce working memory load.
}
