% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ProjectionResults.R
\name{save_ProjectionResults}
\alias{save_ProjectionResults}
\title{Function to save CellType and Pseudotime predictions and UMAP coordinates}
\usage{
save_ProjectionResults(
  query_obj,
  file_name,
  celltype_label = "predicted_CellType",
  celltype_KNNprob_label = "predicted_CellType_prob",
  pseudotime_label = "predicted_Pseudotime"
)
}
\arguments{
\item{query_obj}{Query Seurat object after symphony mapping and celltype classification}

\item{file_name}{File name of csv to store projection annotations and UMAP coordinates}

\item{celltype_label}{Column in query metadata with projected CellType classification.}

\item{celltype_KNNprob_label}{Column in query metadata with KNN probabilities of the projected CellType classification.}

\item{pseudotime_label}{Column in query metadata with predicted Pseudotime values}
}
\value{
Saves a csv file with projected annotations to file_name
}
\description{
Function to save CellType and Pseudotime predictions and UMAP coordinates
}
